backend/requirements.txt
# Core server
aiohttp>=3.11.0
python-socketio>=5.11.0
aiohttp-cors>=0.7.0

# Redis
redis>=5.0.0

# AWS
boto3>=1.34.0

# Auth
PyJWT>=2.8.0

# Environment
python-dotenv>=1.0.0

# Production server (Linux only)
gunicorn>=22.0.0; sys_platform != "win32"
